Adds additional fake data into fixtures (#76)
* Adds additional fake data into fixtures

* Fixed code issue reported by PHP Stan

* Adapted unit tests for new fixtures

+++ src/Repository/UserDonorRepository.php
@@ -62,7 +62,7 @@ class UserDonorRepository extends ServiceEntityRepository
                 'items' => iterator_to_array($paginator),
                 'total' => count($paginator),
                 'current_page' => $page,
-                'total_pages' => ceil(count($paginator) / $limit),
+                'total_pages' => (int) ceil(count($paginator) / $limit),
             ];
         }
